@nineinchnick 2022-09-21
Here's the remaining work required to handle the `nextUri` send in the header: https://github.com/nineinchnick/trino-go-client/compare/concurrent-requests...nineinchnick:trino-go-client:concurrent-requests-header