@nineinchnick 2022-09-22
I need to add a test for a query with a larger number of result rows to make sure we're not dropping any rows.